requirements.txt
Flask
Flask-MySQLdb
Flask-WTF
Flask-Bcrypt
Flask-Login
WTForms